requirements.txt
fastapi[all]
python-decouple
pymongo
pandas
numpy
scikit-learn
openpyxl
psutil
motor